mythic-index/MemoryQuill/story-content/story-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "storylint"
version = "0.1.0"
description = "Local-first story linting for narrative continuity and repository integrity."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Story Lint Team" }]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.6.0",
  "jinja2>=3.1.3",
  "PyYAML>=6.0.1",
  "httpx>=0.27.0",
  "google-adk>=1.0.0",
  "python-dotenv>=1.0.1",
  "rich>=13.7.1",
  "questionary>=2.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]
queue = [
  "celery>=5.3.6",
  "redis>=5.0.1",
]

[project.scripts]
storylint = "storylint_adk.cli:app"
storylint-legacy = "storylint.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["storylint*"]

[tool.setuptools.package-data]
storylint = ["prompts/*.j2", "templates/*.j2"]
storylint_adk = ["prompts/*.j2", "templates/*.j2"]

[tool.pytest.ini_options]
addopts = "-q"
